+++ lib/primitive/lib/Cardano/Wallet/Read/Primitive/Tx/Features/CollateralOutputs.hs
@@ -7,7 +7,7 @@ module Cardano.Wallet.Read.Primitive.Tx.Features.CollateralOutputs
 
 import Prelude
 
-import Cardano.Wallet.Primitive.Convert
+import Cardano.Wallet.Primitive.Ledger.Convert
     ( toWalletTokenBundle
     )
 import Cardano.Wallet.Read.Eras
